Alloxentric-backend/requirements.txt
Flask==2.2.2
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
python-dotenv==0.20.0
gunicorn==20.1.0
eventlet==0.30.2
requests==2.27.1
pymongo==4.2.0
mongoengine==0.24.2
pydub
ffmpeg-python